+++ docs/ChangeLog.md
@@ -1,5 +1,18 @@
 # ChangeLog
 
+## 0.0.4 (Upcoming)
+
+- feat(CargoCmd.profile): MiniStr -> CargoProfile
+- add `CargoProfile`
+
+```rust
+pub enum CargoProfile {
+  Release, // "release"
+  Debug, // "dev"
+  Custom(MiniStr),
+}
+```
+
 ## 0.0.3
 
 - fix(CargoCmd field name): `all_workspaces` -> `all_packages`